Rename to `TollgateAutoConfiguration`
